mysite/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
django-extensions==3.2.3
django-filter==24.2
django-unfold==0.30.0
djangorestframework==3.15.2
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
sympy==1.13.0
thop==0.1.1.post2209072238
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.59
ultralytics-thop==2.0.0
urllib3==2.2.2